box: merge index_read_view.[hc] into read_view.[hc]
The only reason why index_read_view.[hc] was introduced in the first
place was to split the read view implementation between the CE and EE
repositories. Now, when all the read view code resides in the CE
repository, there's no need in it. Let's merge it into the main file
for better code readability.

Follow-up #12912

NO_DOC=code cleanup
NO_TEST=code cleanup
NO_CHANGELOG=code cleanup
